Delete charset.c, code page file support
This does mean that we now only support basic ASCII, since we can no
longer do the correct upper/lower-case mapping for non-ASCII characters.
This is fine; it just means that we are case sensitive in some cases
where we ought to be case insensitive. We can consider adding something
back in the future if it proves essential, but I suspect it won't.