Delete charcnv.c
According to the documentation, this code is for converting filenames
from Windows codepages to Unix ones. Nowadays, Unix filenames should be
UTF-8 anyway so this is kind of redundant. If we want, we can add back
some similar functionality in the future to convert to UTF-8; however,
this code should be redundant either way.