Apply umask when changing permissions
Now that the create mode options have been deleted, we rely on umask to
set sensible permissions for us. For example if our umask is 002 then
created files do not have o+w set. However, when changing (DOS)
permissions we need to do the same thing otherwise bits will be set that
we usually expect to be masked off (when creating new files)