Convert `[fp]strcat` to wrappers around `lstrcat()`
Ideally we should be able to delete `safe_strcat()` entirely along with
the macros, but it does do a couple of things that `lstrcat()` doesn't
(ie. handle `NULL` gracefully and print warning messages when
encountering long strings). We can at least simplify it and change the
calling arguments to match.