Don't run as root; drop privileges, run as nobody
We don't want our server to run as a privileged user since there's no
need for it. Instead, detect if we're running as UID 0 and set all
privileges to the "nobody" user.

It looks like systemd can be configured to run the entire service as an
unprivileged user, however if not running via systemd it may be
desirable to start the program as root so that the privileged sockets
can be opened.