Add Samba 1.8.05 nmbd source files
This is a slightly older version than what we are using for smbd, which
is based on the v1.9 series. However, it looks like lots of features got
added to nmbd between v1.8 and v1.9, and we want something very basic
and simple.

The source tarball was downloaded here:
<https://mirrors.dotsrc.org/samba/old-versions/samba-1.8.05.tar.gz>

All files except those source files needed to build nmbd have been
deleted, to avoid cluttering the Git history.