Rename VAL() macros to allow migration to new ones
There are macros with the same name inside `byteorder.h` but the
behavior is slightly different because they either byteswap or do not
depending on the value of the `NeedSwap` variable. To clean all this up
we need to decide which "version" is needed in each case. To do this
logically in a step-by-step way, let's first rename the old macros so
that the two versions can coexist.