Send browser announcements via broadcast
This ditches the entire mechanism that attempts to look up the browser
master for the workgroup; just broadcast an announcement once a minute
instead. Since we have to broadcast anyway to find the master we might
as well do the simplest thing that works, plus it looks like this may be
what the current version of nmbd is doing anyway.