Reduce update interval field to 60 seconds
We previously had this set to 10 minutes, which meant that servers would
take up to an hour or more in some cases to expire from local browse
masters. Instead, reduce this down to 60 seconds, which is the correct
frequency that we send the announcements (and define a constant that
controls both).