Fix lockups when browsing on Windows 2000
When trying to browse the shares on a Tumba server, the Explorer window
would lock up for two minutes before the shares would finally appear.
This appears to be caused by the Win2k machine trying to open a DCERPC
session on a second TCP/NBSS connection and waiting for a response from
us that we were never sending. We were accepting the session request
even though we didn't know how to interpret the data being sent to us.

Oddly, it seems Win2k only does this when using our own nmbd; when we
run with the Samba nmbd it doesn't try to open a DCERPC session. I'm not
sure why; perhaps something to do with the contents of the browser host
announcements? I attempted changing the contents of those messages to
match Samba but it made no difference ...

Regardless, changing smbd to not accept connections for session types we
don't recognize seems like the better solution here.