Send empty server string in NetServerGetInfo reply
If this is non-empty, IBM Peer (OS/2 Warp) uses it instead of the actual
server name, so the current value ("Tumba 0.2" describing the software
version) becomes very confusing. This does not affect the description
that gets shown on Windows in Network Neighborhood, which comes from the
browser host announcements published by nmbd.