Change the `-W` (workgroup) arg to a hidden option
It should not be necessary to ever actually use this and it does nothing
particularly useful other than set the string that gets included in
various response fields. Having this in the documentation is actually
misleading because it does not affect the workgroup in which the server
appears to other users (in Network Neighborhood etc.) - for that, they
need to configure `nmbd` and its host announcements.