nmbd: Send NetBIOS registration requests on startup
The NBNS spec (RFC 1001) says that we should do this to avoid name
conflicts. The code was actually in the Samba 1.8 source code that we
started from, but was commented out for some reason. This has been
heavily cleaned up and refactored. We don't send browser host
announcements or respond to name requests for our hostname until we have
successfully registered our name with no objections from others.