Abort immediately if we receive a SMBv2 message
Previously, we were returning -1 from this function and entering into a
deadlock state where the client was waiting on a reply that was never
sent. The practical upshot was that `smbclient` would freeze up on
connect if trying to talk to a Tumba server and not configured to
support SMBv1.

Instead, we now abort the server subprocess immediately if any message
is received that is not SMBv1 format, closing the client's connection.
As an additional convenience, also recognize the SMBv2 protocol ID and
log an error message giving the suggested solution if the client is
using Samba.