Drop check on exact value for nm_flags field
As far as I can tell from checking the RFCs, there's no reason or good
rationale for why we should be doing this, and later versions of Samba
don't do it either. This fixes connections from Gnome (and presumably
other software as well, which was sending queries for the server name
with the nm_flags field set to 0, and getting no response.