Don't prefix timestamp for stdout log messages
This is a special case specifically to make Tumba play nicer with
systemd. systemd captures messages that services write to stdout and
commands such as `systemctl status` show the time prefixed. The
`journalctl` command also prefixes the time. We can avoid "doubling" the
timestamp by adding this special case for when `stdout` is being used
for log output.

As an additional point of integration, this also updates the systemd
service file to write log messages to stdout.