Define `STARTUP_ERROR` for fatal startup errors
There are a number of conditions that can lead the program to fail to
start up correctly. These include if it is not possible to open the
server socket, among others. In these cases, instead of writing errors
to the log file, print them instead to `stderr` and exit. This fixes
some previous confusing behavior where the program would just exit
silently if unable to start up and no log output was specified.