Fix status reporting through argv overwrite
On Linux we make use of the hack that allows processes to overwrite
their argv buffer to report their status. However, it seems that the
implementation was wrong: we were overwriting `argv[1]` rather than
`argv[0]`. Recent kernels (?) seem to be truncating to just the first
argument and losing the status information as a result.

This instead changes the implementation to overwrite `argv[0]` which
fixes the bug.