Drop privileges in nmbd, too
As with smbd, nmbd on some systems needs to be started as root so that
it can bind to a privileged port (although this is not necessary on
Linux as it is possible to use security capabilities to grant just that
permission). So let's move the `drop_privileges` function into the
common utility code so that nmbd can use it too.