Raise default log level to `NOTICE()`
This is an attempt at rationalizing some of the log output. There are
certain messages that we do want to print to log, that are part of
normal operation and do not indicate any exceptional condition. In some
cases we were using `WARNING()` to do this, since this was the lowest
log level logged by default.

Clearly this doesn't make any sense - someone connecting to a share for
example is not something that requires a warning condition. Instead,
this raises the default log level to include `NOTICE()` messages, which
we can then use for these kinds of messages.

I've gone through and checked all the existing `NOTICE()` calls and
downgraded many of them to `INFO()` where they don't cover things that
usually deserve any special logging. Similarly, I've also upgraded a
few `INFO()` calls to use `NOTICE()` instead (those used on server
startup/shutdown).