Eliminate `string_{init,free}` functions
There was some optimization here to avoid allocating lots of empty
strings, but I doubt it's worth the added complexity. The only really
useful function here is `string_set`, to simplify the common task of
freeing and replacing a string with a new value. The calls to
`string_init` can be replaced by calls to `checked_strdup`. There was
only one call to `string_free`, which can be replaced by a direct call
to `free`.